requirements.txt
langchain
langchain_community
langchain_groq
langchain_huggingface
python-dotenv
streamlit
pypdf
pypdf2
faiss-cpu